Fixed a wrong function call that was preventing the code to link against the arm gcc toolchain. Function frexpl to frexp

+++ src/fmt_fp.c
@@ -122,7 +122,7 @@ fmt_fp(struct fmt_args *f, long double y, int w, int p, int fl, int t)
     return MAX(w, 3+pl);
   }
 
-  y = frexpl(y, &e2) * 2;
+  y = frexp((double) y, &e2) * 2;
   if (y) e2--;
 
   if ((t|32)=='a') {
